+++ tests/unit/iziModal/MainTest.php
@@ -32,7 +32,7 @@ class MainTest extends \efureev\tests\unit\TestCase
         $jsArray = $view->js[ View::POS_READY ];
         $jsArray = array_flip($jsArray);
 
-        $this->assertTrue(array_key_exists("$('#w1').iziModal();", $jsArray));
+        $this->assertTrue(array_key_exists("$('#w1').iziModal();" . PHP_EOL . "app.modal.init('w1');", $jsArray));
     }
 
     public function testTwo()
@@ -44,7 +44,7 @@ class MainTest extends \efureev\tests\unit\TestCase
         $jsArray = $view->js[ View::POS_READY ];
         $jsArray = array_flip($jsArray);
 
-        $this->assertTrue(array_key_exists("$('#element').iziModal();", $jsArray));
+        $this->assertTrue(array_key_exists("$('#element').iziModal();" . PHP_EOL . "app.modal.init('element');", $jsArray));
     }
 
     public function testThree()
@@ -53,6 +53,39 @@ class MainTest extends \efureev\tests\unit\TestCase
         $input::end();
         $jsArray = $input->view->js[ View::POS_READY ];
         $jsArray = array_flip($jsArray);
-        $this->assertTrue(array_key_exists("$('#element').iziModal();", $jsArray));
+        $this->assertTrue(array_key_exists("$('#element').iziModal();" . PHP_EOL . "app.modal.init('element');", $jsArray));
+    }
+
+    public function testInfo()
+    {
+        $el = 'app-modal';
+        ob_start();
+        ob_implicit_flush(false);
+        IziModalWidget::info($el);
+        $out = ob_get_clean();
+
+        $this->assertEquals($out, '<div id="' . $el . '" class="iziModal" aria-hidden></div>');
+    }
+
+    public function testAlert()
+    {
+        $el = 'app-modal';
+        ob_start();
+        ob_implicit_flush(false);
+        IziModalWidget::alert($el);
+        $out = ob_get_clean();
+
+        $this->assertEquals($out, '<div id="' . $el . '" class="iziModal" aria-hidden></div>');
+    }
+
+    public function testSuccess()
+    {
+        $el = 'app-modal';
+        ob_start();
+        ob_implicit_flush(false);
+        IziModalWidget::success($el);
+        $out = ob_get_clean();
+
+        $this->assertEquals($out, '<div id="' . $el . '" class="iziModal" aria-hidden></div>');
     }
 }