+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ------------------
 * [new] Add JS component - *Implemented*.
 
+
 2016-09-08 - 0.1.1
 ------------------
 * [fix] Change paths - *Fixed*.